Python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatfield"
version = "0.2.0"
description = "Conversational data gathering powered by LLMs - Transform rigid forms into natural conversations"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Chatfield Team"},
]
maintainers = [
    {name = "Chatfield Team"},
]
keywords = [
    "llm",
    "conversation",
    "forms",
    "data-gathering",
    "ai",
    "chatbot",
    "natural-language",
    "validation",
    "langgraph",
    "langchain",
    "openai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]
dependencies = [
    "deepdiff>=6.0.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.29",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.4",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-prebuilt>=0.6.4",
    "langgraph-sdk>=0.2.0",
    "openai>=1.99.6",
    "pydantic>=2.11.7",
    "pydantic_core>=2.33.2",
]

[project.optional-dependencies]
dev = [
    "python_dotenv>=1.1.1",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-describe>=2.2.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "build>=0.10.0",
    "twine>=4.0.0",
]
test = [
    "python_dotenv>=1.1.1",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-describe>=2.2.0",
]
lint = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
notebook = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "notebook>=7.0.0",
    "grandalf>=0.8",
]
visualization = [
    "ipython>=8.0.0",
    "grandalf>=0.8",
]

[project.urls]
Homepage = "https://github.com/yourusername/chatfield"
Documentation = "https://github.com/yourusername/chatfield#readme"
Repository = "https://github.com/yourusername/chatfield"
Issues = "https://github.com/yourusername/chatfield/issues"

[tool.setuptools]
packages = ["chatfield"]

[tool.setuptools.package-data]
chatfield = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/fixtures
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["chatfield"]
skip_gitignore = true
force_single_line = false
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_any_generics = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain.*",
    "langchain_openai.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=chatfield",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "requires_api_key: marks tests that require an API key",
]

[tool.coverage.run]
source = ["chatfield"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if False:",
    "if TYPE_CHECKING:",
]
